// AngularERP/src/app/models/settings/user.model.ts
export class User {
  id: number;
  userTypeId: number;
  roleId:number;
 // userTypeName:string;
  userName: string;
  password: string;
  empId: number;
 // clientId: string;
 // clientName:string;
  compId: number;
 // companyName:string;
 // branchId: number;
 // branchName:string;
 isActive: boolean;
  //empName: string;
  createdDate:Date;
  // createdBy:number;
  // creatorName:string;
  // modifiedDate:Date;
  // modifiedBy:number;
  // modifierName:string;
}